add post visibility
